Created transaction for Order creation

+++ src/main/java/com/example/demo/orderDetails/dto/OrderDetailsDTO.java
@@ -1,10 +1,12 @@
 package com.example.demo.orderDetails.dto;
 
+import java.math.BigDecimal;
+
 public record OrderDetailsDTO(
         Long id,
         Long order_id,
         Long product_id,
         Integer quantity,
-        Integer total
+        BigDecimal total
 ) {
 }
